refactored navbar and added styling
