+++ pages/orgs/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import MyOrgs from '../../components/org/MyOrgs'
-import Navbar from '../../components/project/Navbar'
+import Navbar from '../../components/nav/Navbar'
 import Logo from '../../components/Logo'
 
 export default function Org() {
